# shared/api/openapi.yaml
openapi: 3.1.0
info:
  title: 112 AI Trainer REST API
  version: 0.3.0
  description: Backend scenario API and Voice Gateway API on separate ports.
servers:
- url: http://127.0.0.1:8000/api/v1
  description: Scenario backend
- url: http://127.0.0.1:8001/api/v1
  description: Voice Gateway
security:
- bearerAuth: []
paths:
...
  /auth/status:
    get:
      summary: Status
      operationId: status_api_v1_auth_status_get
      parameters: []
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /auth/bootstrap:
    post:
      summary: Bootstrap
      operationId: bootstrap_api_v1_auth_bootstrap_post
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /auth/login:
    post:
      summary: Login
      operationId: login_api_v1_auth_login_post
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /auth/me:
    get:
      summary: Me
      operationId: me_api_v1_auth_me_get
      parameters:
      - name: X-User-Session
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-User-Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /auth/logout:
    post:
      summary: Logout
      operationId: logout_api_v1_auth_logout_post
      parameters:
      - name: X-User-Session
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-User-Session
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /admin/users:
    get:
      summary: Users
      operationId: users_api_v1_admin_users_get
      parameters:
      - name: X-User-Session
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-User-Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
    post:
      summary: Create
      operationId: create_api_v1_admin_users_post
      parameters:
      - name: X-User-Session
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-User-Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /admin/users/{uid}:
    patch:
      summary: Change Active
      operationId: change_active_api_v1_admin_users__uid__patch
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
          title: Uid
      - name: X-User-Session
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-User-Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserActiveRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    BootstrapRequest:
      properties:
        username:
          type: string
          title: Username
        password:
          type: string
          maxLength: 128
          minLength: 12
          title: Password
        display_name:
          type: string
          maxLength: 100
          minLength: 1
          title: Display Name
      additionalProperties: false
      type: object
      required:
      - username
      - password
      - display_name
      title: BootstrapRequest
...
    CreateUserRequest:
      properties:
        username:
          type: string
          title: Username
        password:
          type: string
          maxLength: 128
          minLength: 12
          title: Password
        display_name:
          type: string
          maxLength: 100
          minLength: 1
          title: Display Name
        role:
          type: string
          enum:
          - teacher
          - student
          title: Role
      additionalProperties: false
      type: object
      required:
      - username
      - password
      - display_name
      - role
      title: CreateUserRequest
    Credentials:
      properties:
        username:
          type: string
          title: Username
        password:
          type: string
          maxLength: 128
          minLength: 12
          title: Password
      additionalProperties: false
      type: object
      required:
      - username
      - password
      title: Credentials
...
    UserActiveRequest:
      properties:
        active:
          type: boolean
          title: Active
      additionalProperties: false
      type: object
      required:
      - active
      title: UserActiveRequest
...